Add findStepWindow function and deprecate findStepsWindow function
Finding the location of the start and end times along a route was
unnecessarily complicated, so we redesigned the window function to find
one time based location at a time.   However, the recursion does not
work currently because it just returns a known leg, and we need to
divide it up.